Pet level should be set properly to wanted value.

+++ L2J_Mobius_10.2_MasterClass/java/org/l2jmobius/gameserver/model/actor/instance/Pet.java
@@ -854,7 +854,7 @@ public class Pet extends Summon
 					return new Pet(template, owner, control);
 				}
 				
-				pet = new Pet(template, owner, control, rset.getByte("level"));
+				pet = new Pet(template, owner, control, rset.getInt("level"));
 				pet._respawned = true;
 				pet.setName(rset.getString("name"));
