+++ L2J_Mobius_Classic_Interlude/java/org/l2jmobius/gameserver/model/PetLevelData.java
@@ -49,7 +49,7 @@ public class PetLevelData
 	public PetLevelData(StatSet set)
 	{
 		_ownerExpTaken = set.getInt("get_exp_type");
-		_petMaxExp = set.getLong("exp");
+		_petMaxExp = (long) set.getDouble("exp");
 		_petMaxHP = set.getFloat("org_hp");
 		_petMaxMP = set.getFloat("org_mp");
 		_petPAtk = set.getFloat("org_pattack");
